+++ route_map/example/lib/custom_model.dart
@@ -1,0 +1,7 @@
+import 'package:route_map/route_map.dart';
+
+@RouteMapArg()
+class CustomModel {
+  String name;
+  CustomModel({required this.name});
+}