+++ route_map/example/lib/detail.dart
@@ -7,8 +7,7 @@ class DetailPage extends StatefulWidget {
   final String id;
   final String name;
   final bool? isShow;
-  const DetailPage(
-      {Key? key, required this.id, required this.name, this.isShow})
+  const DetailPage(this.id, {Key? key, required this.name, this.isShow})
       : super(key: key);
 
   @override