+++ route_map_generator/lib/route_map_generator.dart
@@ -1,4 +1,5 @@
 import 'package:build/build.dart';
+import 'package:route_map_generator/src/route_map_arg_generator.dart';
 import 'package:route_map_generator/src/route_map_config_generator.dart';
 import 'package:route_map_generator/src/route_map_generator.dart';
 import 'package:source_gen/source_gen.dart';
@@ -8,6 +9,14 @@ Builder routeMapBuilder(BuilderOptions options) {
     RouteMapGenerator(),
     formatOutput: (generated) => generated.replaceAll(RegExp(r'//.*|\s'), ''),
     generatedExtension: '.map.json',
+  );
+}
+
+Builder routeMapArgBuilder(BuilderOptions options) {
+  return LibraryBuilder(
+    RouteMapArgGenerator(),
+    formatOutput: (generated) => generated.replaceAll(RegExp(r'//.*|\s'), ''),
+    generatedExtension: '.arg.json',
   );
 }